SistemasDeBigData/UT5-Node Red/Ejercicio5-CarSensor/CarSensorMaria/requirements.txt
redis
python-dotenv
pymongo[srv]
psycopg2-binary
websockets